+++ app/src/components/sections/SingleTeamView/BankTransaction/BankTransactions.vue
@@ -26,10 +26,10 @@
 </template>
 
 <script setup lang="ts">
-import DepositHistory from '@/components/bank-history/DepositHistory.vue'
-import TransferHistory from '@/components/bank-history/TransferHistory.vue'
-import TipsAddressChangedHistory from '@/components/bank-history/TipsAddressChangedHistory.vue'
-import SendToWalletHistory from '@/components/bank-history/SendToWalletHistory.vue'
+import DepositHistory from '@/components/sections/SingleTeamView/BankTransaction/tables/DepositHistory.vue'
+import TransferHistory from '@/components/sections/SingleTeamView/BankTransaction/tables/TransferHistory.vue'
+import TipsAddressChangedHistory from '@/components/sections/SingleTeamView/BankTransaction/tables/TipsAddressChangedHistory.vue'
+import SendToWalletHistory from '@/components/sections/SingleTeamView/BankTransaction/tables/SendToWalletHistory.vue'
 import AccordionComponent from '@/components/AccordionComponent.vue'
 import { ref } from 'vue'
