+++ app/src/components/sections/SingleTeamView/BankTransaction/tables/__tests__/DepositHistory.spec.ts
@@ -1,6 +1,6 @@
 import { describe, it, expect, beforeEach, vi, type Mock } from 'vitest'
 import { VueWrapper, mount } from '@vue/test-utils'
-import SendToWalletHistory from '@/components/bank-history/SendToWalletHistory.vue'
+import DepositHistory from '@/components/sections/SingleTeamView/BankTransaction/tables/DepositHistory.vue'
 import SkeletonLoading from '@/components/SkeletonLoading.vue'
 import type { EventResult } from '@/types'
 import type { Result } from 'ethers'
@@ -9,15 +9,15 @@ import { useBankEvents } from '@/composables/bank'
 import { createTestingPinia } from '@pinia/testing'
 import { NETWORK } from '@/constant'
 
-const sendToWalletEvents: EventResult[] = [
+const depositEvents: EventResult[] = [
   {
     txHash: '0x1',
-    data: ['0xOwner', ['0xMember1', '0xMember2'], '1000000000000000000'] as Result, // 1 ETH
+    data: ['0xDepositor1', '1000000000000000000'] as Result, // 1 ETH
     date: '2024-06-25'
   },
   {
     txHash: '0x2',
-    data: ['0xOwner', ['0xMember1', '0xMember2'], '2000000000000000000'] as Result, // 2 ETH
+    data: ['0xDepositor2', '2000000000000000000'] as Result, // 2 ETH
     date: '2024-06-26'
   }
 ]
@@ -37,18 +37,18 @@ vi.mock('@/stores/useToastStore', () => ({
 
 vi.mock('@/composables/bank', () => ({
   useBankEvents: vi.fn().mockImplementation(() => ({
-    getEvents: vi.fn().mockReturnValue(sendToWalletEvents),
+    getEvents: vi.fn().mockReturnValue(depositEvents),
     loading: ref(false),
-    events: ref(sendToWalletEvents),
+    events: ref(depositEvents),
     error: ref(null)
   }))
 }))
 
-describe('SendToWalletHistory', () => {
+describe('DepositHistory', () => {
   let wrapper: VueWrapper
 
   beforeEach(() => {
-    wrapper = mount(SendToWalletHistory, {
+    wrapper = mount(DepositHistory, {
       props: {
         bankAddress: '0x123'
       },
@@ -67,27 +67,25 @@ describe('SendToWalletHistory', () => {
       expect(wrapper.find('tbody tr').exists()).toBe(true)
 
       // table header
-      const header = ['No', 'Owner Address', 'Member Addresses', 'Total Amount', 'Date']
-      expect(wrapper.findAll('th').length).toBe(5)
+      const header = ['No', 'Depositor', 'Amount', 'Date']
+      expect(wrapper.findAll('th').length).toBe(4)
       expect(wrapper.findAll('th').forEach((th, index) => expect(th.text()).toBe(header[index])))
 
       // table body
-      expect(wrapper.findAll('td').length).toBe(sendToWalletEvents.length * header.length)
+      expect(wrapper.findAll('td').length).toBe(depositEvents.length * header.length)
     })
 
     it('renders table body correctly', () => {
       const tableData = wrapper.findAll('td')
       const no = tableData[0].text()
-      const ownerAddress = tableData[1].text()
-      const memberAddresses = tableData[2].text()
-      const totalAmount = tableData[3].text()
-      const date = tableData[4].text()
+      const depositor = tableData[1].text()
+      const amount = tableData[2].text()
+      const date = tableData[3].text()
 
       expect(no).toEqual('1')
-      expect(ownerAddress).toEqual(sendToWalletEvents[0].data[0])
-      expect(memberAddresses).toEqual(sendToWalletEvents[0].data[1].join(''))
-      expect(totalAmount).toEqual('1 SepoliaETH')
-      expect(date).toEqual(sendToWalletEvents[0].date)
+      expect(depositor).toEqual(depositEvents[0].data[0])
+      expect(amount).toEqual('1 SepoliaETH')
+      expect(date).toEqual(depositEvents[0].date)
     })
 
     it('renders skeleton loading if loading', () => {
@@ -98,7 +96,7 @@ describe('SendToWalletHistory', () => {
         error: ref(null)
       }))
 
-      const wrapper = mount(SendToWalletHistory, {
+      const wrapper = mount(DepositHistory, {
         props: {
           bankAddress: '0x123'
         }
@@ -114,15 +112,15 @@ describe('SendToWalletHistory', () => {
         error: ref(null)
       }))
 
-      const wrapper = mount(SendToWalletHistory, {
+      const wrapper = mount(DepositHistory, {
         props: {
           bankAddress: '0x123'
         }
       })
       const emtpyRow = wrapper.find('tr td.text-center.font-bold.text-lg')
       expect(emtpyRow.exists()).toBe(true)
-      expect(emtpyRow.text()).toBe('No send to wallet history')
-      expect(emtpyRow.attributes('colspan')).toBe('5')
+      expect(emtpyRow.text()).toBe('No Deposit transactions')
+      expect(emtpyRow.attributes('colspan')).toBe('4')
     })
   })
 
@@ -136,7 +134,7 @@ describe('SendToWalletHistory', () => {
         error: ref(null)
       }))
 
-      mount(SendToWalletHistory, {
+      mount(DepositHistory, {
         props: {
           bankAddress: '0x123'
         }
